New id system
Now pieces are defined with up to 2 ids (for white and black) instead of always using uppercase for white and lowercase for black.
This means that any unicode character can be used as id, not only uppercase letters.
This is also a requisite for implementing fully generic FEN parsing, which in the future might replace the current Vec<PiecePlacement> (but first there needs to be a way to encode which pieces have moved in the FEN string).